[RISCV][VP] Introduce vp saturating addition/substraction and RISC-V support.
This patch also pick the MatchContext framework from DAGCombiner to an indiviual
header file to make the framework be used from other files in
llvm/lib/CodeGen/SelectionDAG/.
